requirements.txt
streamlit
openai
faiss-cpu
PyMuPDF
sentence-transformers
transformers
python-dotenv
